# shared/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/exchange": {
      "get": {
        "operationId": "AuthController_exchange",
        "parameters": [
          {
            "name": "t",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to post-login path or login-failed"
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "userId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "nullable": true
                    },
                    "displayName": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "channels": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "channel": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string",
                            "nullable": true
                          },
                          "displayName": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": ["channel", "username", "displayName"]
                      }
                    }
                  },
                  "required": [
                    "userId",
                    "email",
                    "displayName",
                    "avatarUrl",
                    "channels"
                  ]
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["auth"]
      },
      "delete": {
        "operationId": "AuthController_deleteMe",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Account permanently deleted"
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/link/telegram/start": {
      "post": {
        "operationId": "AuthController_linkTelegramStart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": ["url"]
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/link/telegram": {
      "delete": {
        "operationId": "AuthController_unlinkTelegram",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Telegram identity unlinked"
          },
          "401": {
            "description": ""
          },
          "409": {
            "description": "Cannot remove the last remaining login method"
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/link/google": {
      "delete": {
        "operationId": "AuthController_unlinkGoogle",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Google identity unlinked"
          },
          "401": {
            "description": ""
          },
          "409": {
            "description": "Cannot remove the last remaining login method"
          }
        },
        "tags": ["auth"]
      }
    },
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailAuthDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Registered and session set"
          },
          "409": {
            "description": "Email already registered"
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailAuthDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Authenticated and session set"
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["auth"]
      }
    },
...
    "/auth/google/callback": {
      "get": {
        "operationId": "AuthController_googleCallback",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to post-login path or error"
          }
        },
        "tags": ["auth"]
      }
    },
...
    "/auth/telegram/webapp": {
      "post": {
        "operationId": "AuthController_telegramWebApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramWebAppAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "userId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "nullable": true
                    },
                    "displayName": {
                      "type": "string",
                      "nullable": true
                    },
                    "avatarUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "channels": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "channel": {
                            "type": "string"
                          },
                          "username": {
                            "type": "string",
                            "nullable": true
                          },
                          "displayName": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": ["channel", "username", "displayName"]
                      }
                    }
                  },
                  "required": [
                    "userId",
                    "email",
                    "displayName",
                    "avatarUrl",
                    "channels"
                  ]
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        },
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "Stickup API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid"
      }
    },
    "schemas": {
...
      "EmailAuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": ["email", "password"]
      },
      "TelegramWebAppAuthDto": {
        "type": "object",
        "properties": {
          "initData": {
            "type": "string",
            "description": "Raw initData string from Telegram WebApp",
            "maxLength": 4096
          }
        },
        "required": ["initData"]
      }
    }
  }
}